+++ .env
@@ -1,6 +1,4 @@
     GCLOUD_PROJECT_ID=crud-happens
-
- GCLOUD_APPLICATION_CREDENTIALS=./ team -crud-happens-backend/services/firebase-key.json
- 
- GCLOUD_STORAGE_BUCKET_URL= crud-happens.appspot.com
+ GCLOUD_APPLICATION_CREDENTIALS=./services/firebase-key.json
+  GCLOUD_STORAGE_BUCKET_URL= crud-happens.appspot.com
