+++ package.json
@@ -24,6 +24,7 @@
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
+    "firebase": "^8.6.8",
     "mongoose": "^5.12.13",
     "multer": "^1.4.2",
     "nodemon": "^2.0.7"
